refactor(app): moved hooks calling view functions into `fetch` dir
